Improved force field handling in trajectory files
Force fields that are not part of MMTK itself could not be used easily
in systems from which trajectories are generated. The system description
in the trajectory would systematically raise an exception because MMTK
didn't know from where to get the force field. With this modification,
the full path to the force field class is recorded in the trajectory,
and the code that reconstructs the system when the trajectory is read
imports the module(s) in order to create the force field object.
The only requirement is that the module containing the force field
class must be importable, i.e. be located on sys.path.